# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is Swagger UI environment generated for the TMF Product Catalog Management specification",
    "version" : "2.0",
    "title" : "Product Catalog Management"
  },
  "host" : "env-0693795.jelastic.servint.net",
  "basePath" : "/DSProductCatalog/api/catalogManagement/v2",
  "schemes" : [ "http" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
...
    "/productOffering" : {
      "get" : {
        "tags" : [ "productOffering" ],
        "summary" : "List or find 'ProductOffering' objects",
        "operationId" : "listProductOffering",
        "parameters" : [ {
          "name" : "fields",
          "in" : "query",
          "description" : "Comma separated properties to display in response",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ProductOffering"
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "productOffering" ],
        "summary" : "Creates a 'ProductOffering'",
        "operationId" : "createProductOffering",
        "parameters" : [ {
          "in" : "body",
          "name" : "productOffering",
          "description" : "The Product Offering to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ProductOffering_Create"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created",
            "schema" : {
              "$ref" : "#/definitions/ProductOffering"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/productOffering/{id}" : {
      "get" : {
        "tags" : [ "productOffering" ],
        "summary" : "Retrieves a 'ProductOffering' by Id",
        "operationId" : "retrieveProductOffering",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Product Offering",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "fields",
          "in" : "query",
          "description" : "Comma separated properties to display in response",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ProductOffering"
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "productOffering" ],
        "summary" : "Updates a 'ProductOffering' by Id",
        "operationId" : "updateProductOffering",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Product Offering",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "productOffering",
          "description" : "The Product Offering to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ProductOffering_Update"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/ProductOffering"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "productOffering" ],
        "summary" : "Updates partially a 'ProductOffering' by Id",
        "operationId" : "updateProductOffering",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Product Offering",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "productOffering",
          "description" : "The Product Offering to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ProductOffering_Update"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/ProductOffering"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/productSpecification" : {
      "get" : {
        "tags" : [ "productSpecification" ],
        "summary" : "List or find 'ProductSpecification' objects",
        "operationId" : "listProductSpecification",
        "parameters" : [ {
          "name" : "fields",
          "in" : "query",
          "description" : "Comma separated properties to display in response",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ProductSpecification"
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "productSpecification" ],
        "summary" : "Creates a 'ProductSpecification'",
        "operationId" : "createProductSpecification",
        "parameters" : [ {
          "in" : "body",
          "name" : "productSpecification",
          "description" : "The Product Specification to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ProductSpecification_Create"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created",
            "schema" : {
              "$ref" : "#/definitions/ProductSpecification"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/productSpecification/{id}" : {
      "get" : {
        "tags" : [ "productSpecification" ],
        "summary" : "Retrieves a 'ProductSpecification' by Id",
        "operationId" : "retrieveProductSpecification",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Product Specification",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "fields",
          "in" : "query",
          "description" : "Comma separated properties to display in response",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ProductSpecification"
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "productSpecification" ],
        "summary" : "Updates a 'ProductSpecification' by Id",
        "operationId" : "updateProductSpecification",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Product Specification",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "productSpecification",
          "description" : "The Product Specification to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ProductSpecification_Update"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/ProductSpecification"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "productSpecification" ],
        "summary" : "Updates partially a 'ProductSpecification' by Id",
        "operationId" : "updateProductSpecification",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Product Specification",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "productSpecification",
          "description" : "The Product Specification to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/ProductSpecification_Update"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/ProductSpecification"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ProductOffering" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the productOffering"
        },
        "href" : {
          "type" : "string",
          "description" : "Reference of the ProductOffering"
        },
        "version" : {
          "type" : "string",
          "description" : "ProductOffering version"
        },
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date and time of the last update"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the productOffering"
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the productOffering"
        },
        "isBundle" : {
          "type" : "boolean",
          "description" : "isBundle determines whether a productOffering represents a single productOffering (false), or a bundle of productOfferings (true)."
        },
        "lifecycleStatus" : {
          "type" : "string",
          "description" : "Used to indicate the current lifecycle status"
        },
        "validFor" : {
          "description" : "The period for which the productOffering is valid",
          "$ref" : "#/definitions/TimePeriod"
        },
        "place" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PlaceRef"
          }
        },
        "serviceLevelAgreement" : {
          "$ref" : "#/definitions/ServiceLevelAgreementRef"
        },
        "productSpecification" : {
          "$ref" : "#/definitions/ProductSpecificationRef"
        },
        "channel" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ChannelRef"
          }
        },
        "serviceCandidate" : {
          "$ref" : "#/definitions/ServiceCandidateRef"
        },
        "category" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CategoryRef"
          }
        },
        "resourceCandidate" : {
          "$ref" : "#/definitions/ResourceCandidateRef"
        },
        "productOfferingPrice" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductOfferingPrice"
          }
        },
        "bundledProductOffering" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/BundledProductOffering"
          }
        },
        "productOfferingTerm" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductOfferingTerm"
          }
        }
      },
      "description" : "Represents entities that are orderable from the provider of the catalog, this resource includes pricing information."
    },
    "ProductOffering_Create" : {
      "type" : "object",
      "required" : [ "bundledProductOffering", "name", "productOfferingPrice", "productSpecification" ],
      "properties" : {
        "version" : {
          "type" : "string",
          "description" : "ProductOffering version"
        },
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date and time of the last update"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the productOffering"
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the productOffering"
        },
        "isBundle" : {
          "type" : "boolean",
          "description" : "isBundle determines whether a productOffering represents a single productOffering (false), or a bundle of productOfferings (true)."
        },
        "lifecycleStatus" : {
          "type" : "string",
          "description" : "Used to indicate the current lifecycle status"
        },
        "validFor" : {
          "description" : "The period for which the productOffering is valid",
          "$ref" : "#/definitions/TimePeriod"
        },
        "place" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PlaceRef"
          }
        },
        "serviceLevelAgreement" : {
          "$ref" : "#/definitions/ServiceLevelAgreementRef"
        },
        "productSpecification" : {
          "$ref" : "#/definitions/ProductSpecificationRef"
        },
        "channel" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ChannelRef"
          }
        },
        "serviceCandidate" : {
          "$ref" : "#/definitions/ServiceCandidateRef"
        },
        "category" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CategoryRef"
          }
        },
        "resourceCandidate" : {
          "$ref" : "#/definitions/ResourceCandidateRef"
        },
        "productOfferingPrice" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductOfferingPrice"
          }
        },
        "bundledProductOffering" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/BundledProductOffering"
          }
        },
        "productOfferingTerm" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductOfferingTerm"
          }
        }
      },
      "description" : "Represents entities that are orderable from the provider of the catalog, this resource includes pricing information.\nSkipped properties: id,href"
    },
    "ProductOffering_Update" : {
      "type" : "object",
      "properties" : {
        "version" : {
          "type" : "string",
          "description" : "ProductOffering version"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the productOffering"
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the productOffering"
        },
        "isBundle" : {
          "type" : "boolean",
          "description" : "isBundle determines whether a productOffering represents a single productOffering (false), or a bundle of productOfferings (true)."
        },
        "lifecycleStatus" : {
          "type" : "string",
          "description" : "Used to indicate the current lifecycle status"
        },
        "validFor" : {
          "description" : "The period for which the productOffering is valid",
          "$ref" : "#/definitions/TimePeriod"
        },
        "place" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/PlaceRef"
          }
        },
        "serviceLevelAgreement" : {
          "$ref" : "#/definitions/ServiceLevelAgreementRef"
        },
        "productSpecification" : {
          "$ref" : "#/definitions/ProductSpecificationRef"
        },
        "channel" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ChannelRef"
          }
        },
        "serviceCandidate" : {
          "$ref" : "#/definitions/ServiceCandidateRef"
        },
        "category" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/CategoryRef"
          }
        },
        "resourceCandidate" : {
          "$ref" : "#/definitions/ResourceCandidateRef"
        },
        "productOfferingPrice" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductOfferingPrice"
          }
        },
        "bundledProductOffering" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/BundledProductOffering"
          }
        },
        "productOfferingTerm" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductOfferingTerm"
          }
        }
      },
      "description" : "Represents entities that are orderable from the provider of the catalog, this resource includes pricing information.\nSkipped properties: id,href,id,lastUpdate"
    },
    "ProductSpecification" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the product specification"
        },
        "href" : {
          "type" : "string",
          "description" : "Reference of the product specification"
        },
        "version" : {
          "type" : "string",
          "description" : "Product specification version"
        },
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date and time of the last update"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the product specification"
        },
        "description" : {
          "type" : "string",
          "description" : "A narrative that explains in detail what the product specification is"
        },
        "isBundle" : {
          "type" : "boolean",
          "description" : "isBundle determines whether a productSpecification represents a single productSpecification (false), or a bundle of productSpecification (true)."
        },
        "brand" : {
          "type" : "string",
          "description" : "The manufacturer or trademark of the specification"
        },
        "lifecycleStatus" : {
          "type" : "string",
          "description" : "Used to indicate the current lifecycle status"
        },
        "validFor" : {
          "description" : "The period for which the product specification is valid",
          "$ref" : "#/definitions/TimePeriod"
        },
        "productNumber" : {
          "type" : "string",
          "description" : "An identification number assigned to uniquely identity the specification"
        },
        "relatedParty" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/RelatedPartyRef"
          }
        },
        "productSpecCharacteristic" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductSpecCharacteristic"
          }
        },
        "serviceSpecification" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ServiceSpecificationRef"
          }
        },
        "productSpecificationRelationship" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ProductSpecificationRelationship"
          }
        },
        "resourceSpecification" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ResourceSpecificationRef"
          }
        },
        "attachment" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Attachment"
          }
        },
        "bundledProductSpecification" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/BundledProductSpecification"
          }
        }
      },
      "description" : "Is a detailed description of a tangible or intangible object made available externally in the form of a ProductOffering to customers or other parties playing a party role."
    },
...
    "TimePeriod" : {
      "type" : "object",
      "properties" : {
        "startDateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "An instant of time, starting at the TimePeriod"
        },
        "endDateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "An instant of time, ending at the TimePeriod."
        }
      },
      "description" : "A base / value business entity used to represent a period of time between two timepoints."
    },
...
    "PlaceRef" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the place"
        },
        "href" : {
          "type" : "string",
          "description" : "Unique reference of the place"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the place"
        }
      },
      "description" : "Place reference. Place defines the places where the product offerings are sold."
    },
...
    "Error" : {
      "required" : [ "code", "message" ],
      "properties" : {
        "code" : {
          "type" : "integer",
          "description" : "An integer coding the error type. This is given to caller so he can translate them if required."
        },
        "message" : {
          "type" : "string",
          "description" : "A short localized string that describes the error."
        },
        "description" : {
          "type" : "string",
          "description" : "(optional) A long localized error description if needed. It can contain precise information about which parameter is missing, or what are the identifier acceptable values."
        },
        "infoURL" : {
          "type" : "string",
          "description" : "(optional) A URL to online documentation that provides more information about the error."
        }
      }
    }
  }
}